Add files via upload
First, create the index.html file and style.css file. After you created those files then start the HTML part of the cursor element in the index.html file. When the HTML part is complete, to start the styling part with the cursor element animation link the style.css file to the index.html file in between <head></head> tags. Then go to the style.css file and do the styling part of the cursor element animation as you can see in the video. After the styling part complete then go to the index.html file for the javascript part. In the index.html file before the body close tag open <script></script> tags and complete the javascript part inside that tags as you can see in the project. So that's all. Hope this video will be helpful to create awesome cursor effects on a website.